Extract getPurchaseEvents
